% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinynotyf.R
\docType{data}
\name{icons}
\alias{icons}
\title{Material Icon Names.}
\format{A list with 932 elements.}
\usage{
icons
}
\description{
A listing containing the names of many (if not all) material icon names.
}
\keyword{datasets}
