% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinynotyf.R
\name{shinynotyf}
\alias{shinynotyf}
\title{Display a notification message (notyf) in Shiny}
\usage{
shinynotyf(message, type = "custom", duration = NULL,
  color = "#0a84ff", ripple = FALSE, icon = "face")
}
\arguments{
\item{message}{Message to display in the notification.}

\item{type}{Type of notification to display, one of custom, success, or
error. Defaults to custom.}

\item{duration}{Duration of the notification.}

\item{color}{Backgroun color of the notification.}

\item{ripple}{Enables ripple animation, defaults to FALSE.}

\item{icon}{Material icon to be used, defaults to 'face'.}
}
\description{
A notyf is a simple notification that contains a message and an icon. The
message, icon, duration, color, and ripple effect can all be modified. There
is an additional parameter \emph{type} that can be used as a shortcut to
display a custom, success, or error notification.
}
