% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinynotyf.R
\name{use_shinynotyf}
\alias{use_shinynotyf}
\title{Set up a Shiny app to use shinynotyf}
\usage{
use_shinynotyf()
}
\value{
Scripts that \code{shinynotyf} requires that are automatically
inserted to the app's \code{<head>} tag.
}
\description{
This function must be called from a Shiny app's UI in order for the
\code{\link[shinynotyf]{shinynotyf}} function to work.\cr\cr
You can call \code{use_shinynotyf()} from anywhere inside the UI.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinynotyf)

  shinyApp(
    ui = fluidPage(
      # Set up shinynotyf
      use_shinynotyf(),
      actionButton("btn", "Click me")
    ),
    server = function(input, output) {
      observeEvent(input$btn, {
        # Show a simple notification
        shinynotyf(message = "You did it!", type = "success")
      })
    }
  )
}
}
\seealso{
\code{\link[shinynotyf]{shinynotyf}}
}
